Fixed all remaining issues with teams.
When players changed their team in game, only the host saw this change
in lobby. This has been fixed.
When opening the select team menu in game, the start menu would close.
Also added team colors to chat.
And fixed an issue where setup pane data was not displayed returning to
lobby.